requirements.txt
pandas
faiss-cpu
sentence-transformers
numpy
requests
beautifulsoup4

